# perf: cache candidates seperate from their sorting