--- a/crates/libsolv_rs/Cargo.toml
+++ b/crates/libsolv_rs/Cargo.toml
@@ -11,6 +11,7 @@ license.workspace = true
 readme.workspace = true
 
 [dependencies]
+ahash = "0.8.3"
 itertools = "0.11.0"
 petgraph = "0.6.3"
 rattler_conda_types = { version = "0.5.0", path = "../rattler_conda_types" }