--- a/crates/libsolv_rs/src/conda_util.rs
+++ b/crates/libsolv_rs/src/conda_util.rs
@@ -9,6 +9,7 @@ use std::cmp::Ordering;
 use std::collections::HashMap;
 
 /// Returns the order of two candidates based on the order used by conda.
+#[allow(clippy::too_many_arguments)]
 pub(crate) fn compare_candidates(
     a: SolvableId,
     b: SolvableId,
@@ -17,6 +18,7 @@ pub(crate) fn compare_candidates(
     packages_by_name: &Mapping<NameId, Vec<SolvableId>>,
     match_specs: &Arena<MatchSpecId, MatchSpec>,
     match_spec_to_candidates: &Mapping<MatchSpecId, OnceCell<Vec<SolvableId>>>,
+    match_spec_highest_version: &Mapping<MatchSpecId, OnceCell<Option<(Version, bool)>>>,
 ) -> Ordering {
     let a_solvable = solvables[a].package();
     let b_solvable = solvables[b].package();
@@ -81,6 +83,7 @@ pub(crate) fn compare_candidates(
                 packages_by_name,
                 match_specs,
                 match_spec_to_candidates,
+                match_spec_highest_version,
             );
             let highest_b = find_highest_version(
                 *b_spec_id,
@@ -89,6 +92,7 @@ pub(crate) fn compare_candidates(
                 packages_by_name,
                 match_specs,
                 match_spec_to_candidates,
+                match_spec_highest_version,
             );
 
             // Skip version if no package is selected by either spec
@@ -140,43 +144,49 @@ pub(crate) fn find_highest_version(
     packages_by_name: &Mapping<NameId, Vec<SolvableId>>,
     match_specs: &Arena<MatchSpecId, MatchSpec>,
     match_spec_to_candidates: &Mapping<MatchSpecId, OnceCell<Vec<SolvableId>>>,
+    match_spec_highest_version: &Mapping<MatchSpecId, OnceCell<Option<(Version, bool)>>>,
 ) -> Option<(Version, bool)> {
-    let candidates = find_candidates(
-        match_spec_id,
-        match_specs,
-        interned_strings,
-        packages_by_name,
-        solvables,
-        match_spec_to_candidates,
-    );
+    match_spec_highest_version[match_spec_id]
+        .get_or_init(|| {
+            let candidates = find_candidates(
+                match_spec_id,
+                match_specs,
+                interned_strings,
+                packages_by_name,
+                solvables,
+                match_spec_to_candidates,
+            );
 
-    candidates
-        .iter()
-        .map(|id| &solvables[*id].package().record)
-        .fold(None, |init, record| {
-            Some(init.map_or_else(
-                || {
-                    (
-                        record.version.version().clone(),
-                        !record.track_features.is_empty(),
-                    )
-                },
-                |(version, has_tracked_features)| {
-                    (
-                        version.max(record.version.version().clone()),
-                        has_tracked_features && record.track_features.is_empty(),
-                    )
-                },
-            ))
+            candidates
+                .iter()
+                .map(|id| &solvables[*id].package().record)
+                .fold(None, |init, record| {
+                    Some(init.map_or_else(
+                        || {
+                            (
+                                record.version.version().clone(),
+                                !record.track_features.is_empty(),
+                            )
+                        },
+                        |(version, has_tracked_features)| {
+                            (
+                                version.max(record.version.version().clone()),
+                                has_tracked_features && record.track_features.is_empty(),
+                            )
+                        },
+                    ))
+                })
         })
+        .as_ref()
+        .map(|(version, has_tracked_features)| (version.clone(), *has_tracked_features))
 }
 
-pub(crate) fn find_candidates<'a, 'b>(
+pub(crate) fn find_candidates<'b>(
     match_spec_id: MatchSpecId,
     match_specs: &Arena<MatchSpecId, MatchSpec>,
     names_to_ids: &HashMap<String, NameId>,
     packages_by_name: &Mapping<NameId, Vec<SolvableId>>,
-    solvables: &Arena<SolvableId, Solvable<'a>>,
+    solvables: &Arena<SolvableId, Solvable>,
     match_spec_to_candidates: &'b Mapping<MatchSpecId, OnceCell<Vec<SolvableId>>>,
 ) -> &'b Vec<SolvableId> {
     match_spec_to_candidates[match_spec_id].get_or_init(|| {