--- a/crates/libsolv_rs/src/problem.rs
+++ b/crates/libsolv_rs/src/problem.rs
@@ -52,7 +52,7 @@ impl Problem {
                 Clause::Requires(package_id, match_spec_id) => {
                     let package_node = Self::add_node(&mut graph, &mut nodes, package_id);
 
-                    let candidates = &solver.pool().match_spec_to_candidates[match_spec_id];
+                    let candidates = &solver.pool().match_spec_to_sorted_candidates[match_spec_id];
                     if candidates.is_empty() {
                         tracing::info!(
                             "{package_id:?} requires {match_spec_id:?}, which has no candidates"