--- a/crates/libsolv_rs/src/solver/mod.rs
+++ b/crates/libsolv_rs/src/solver/mod.rs
@@ -9,10 +9,10 @@ use crate::solve_jobs::SolveJobs;
 use crate::transaction::Transaction;
 use std::cell::OnceCell;
 
+use ahash::{AHashMap, AHashSet};
 use itertools::Itertools;
 use rattler_conda_types::MatchSpec;
 use std::collections::{HashMap, HashSet};
-use ahash::{AHashMap, AHashSet};
 
 use clause::{Clause, ClauseState, Literal};
 use decision::Decision;