Improve performance of the solver by caching the candidates separately from the sorted candidates. The candidates were also computed to compare different solvables.

With these improvements libsolv-rs beats the C impl in every case. (on my machine..)

|                                    | libsolv | libsolv-rs |
|------------------------------------|---------|------------|
| python=3.9                         | 13.267ms | **8.60ms**    |
| xtensor, xsimd                     | 9.7439ms | **4.15ms**    |
| tensorflow                           | 1.3159s | **0.753s**    |
| quetz                                   | 3.09s | **2.38s**    |
| tensorboard=2.1.1, grpc-cpp=1.39.1 | 1.5399s | **0.241s**    |