@ruben-arts 2023-07-06
req | libsolv | libsolv-rs
-- | -- | --
python=3.9 | 8.9680 ms | 4.1932 ms
xtensor, xsimd | 7.0295 ms | 2.3236 9ms
tensorflow | 931.89 ms | 529.56 ms
quetz | 1.8615 s | 1.6781 s
tensorboard=2.1.1, grpc-cpp=1.39.1 | 698.74 ms | 202.47 ms