@wolfv 2023-07-07
Can you explain why AHashMap is faster? I looked at the crate briefly but it wasn't obvious to me why there would be a performance improvement.

Everything else looks great!