@wolfv 2023-07-07
One other thing I noticed (and maybe that's related) is the order of the packages printed to-be-installed became non-deterministic.

I am not sure that it's what we want. We might even want the opposite and order the matchspecs initially to get results independent of spec-ordering.