@baszalmstra 2023-07-07
> Can you explain why AHashMap is faster? I looked at the crate briefly but it wasn't obvious to me why there would be a performance improvement.

`AHashMap` should be faster for almost all cases but for me the results have been not been enormous. See https://github.com/tkaitchuck/aHash/blob/master/compare/readme.md#fxhash why we should use ahash.

> One other thing I noticed (and maybe that's related) is the order of the packages printed to-be-installed became non-deterministic.

I think it should be left up to the user to determine how they want to sort the packages. Given that it is now (especially with your changes) relatively easy to generate a topological sort from the result I say we leave this as is.

> I am not sure that it's what we want. We might even want the opposite and order the matchspecs initially to get results independent of spec-ordering.

Yeah, I also find it not very user friendly that `a b` might resolve differently than `b a`. However, Im not sure if this is something we should implement in the solver or whatever thats a useful artifact of the solver. Everyone using rattler is of course free to sort the input specs if they want.