@baszalmstra 2023-07-07
> If I understood the AHashMap correctly, then it uses a source of randomness for the hashing. I guess I am wondering if that source of randomness could alter the resolutions.

As far as I know, the only place we use this HashMap is to lookup keys to values. I don't think we are iterating over the hashmaps anywhere which could indeed cause that behavior. 

Do you see different results on every run?

> The thing is that the order of evaluation of the packages influences the result (e.g. when you ask for xtensor xsimd you get something else from xsimd xtensor.

That was already the case right? Or does this PR change that?