@wolfv 2023-07-07
The order of prints is already different every time in main. So let's forget about it for now.